Exception handling - delete

+++ src/main/java/com/marques/springproject/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.marques.springproject.entities.User;
 import com.marques.springproject.repositories.UserRepository;
+import com.marques.springproject.services.exceptions.DatabaseException;
 import com.marques.springproject.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +32,21 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        try{
+            if (repository.existsById(id)){
+                repository.deleteById(id);
+            }else {
+                throw new ResourceNotFoundException(id);
+
+            }
+
+        }catch (ResourceNotFoundException e){
+            throw new ResourceNotFoundException("Id : " + id + " does not exist");
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
+
     }
 
     public User update(Long id, User obj){
